+++ src/main/java/subway/station/controller/dto/StationResponse.java
@@ -1,4 +1,4 @@
-package subway;
+package subway.station.controller.dto;
 
 public class StationResponse {
     private Long id;
