+++ src/main/java/subway/station/domain/Station.java
@@ -1,4 +1,4 @@
-package subway;
+package subway.station.domain;
 
 import javax.persistence.*;
