+++ src/test/java/subway/StationAcceptanceTest.java
@@ -15,6 +15,7 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.test.annotation.DirtiesContext;
 import org.springframework.test.annotation.DirtiesContext.ClassMode;
+import subway.station.controller.dto.StationResponse;
 
 @DisplayName("지하철역 관련 기능")
 @DirtiesContext(classMode = ClassMode.BEFORE_EACH_TEST_METHOD)
